Fixed markdown formatting in 4.0authentification.md
I was referred to this page after asking a question on reddit and noticed that the markdown for the headers was not formatted properly. There was no space after the #s, so they were showing up in the actual displayed file. I don’t know if this occurs in any of the other files in this repository, but I thought I’d fix this one real quick. Hope it helps.